refactor(campaignActions): rename section back to CampaignActions
