+++ src/styles/index.scss
@@ -81,4 +81,4 @@ $header-height-md: 8rem;
 @import "../sections/CTA/style";
 @import "../components/Footer/style";
 @import "../components/Header/style";
-@import "../sections/Campaign/style";
+@import "../sections/CampaignActions/style";
