+++ src/templates/actions-page.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Layout from "../components/Layout";
-import Campaign from "../sections/Campaign";
+import Campaign from "../sections/CampaignActions";
 
 // TODO: this data needs to be pulled from our SSO service
 const tempFakeUser = {
